Fix issue where uninitialized result would be stored with previous values
Whenever an aircraft lookup failed because there was not an entry in
basestation.sqb, the result would keep the values from the previous lookup,
therefore storing duplicated data.

To fix this, the result array is now initialized to "N/A" strings before
performing the actual lookup. If there are no entries for a particular
aircraft, then instead of using the previous values, the aircraft is stored
with "N/A" strings in each field.